focuschallenge/main/api/serializers.py
from rest_framework import serializers
from main.models import Employee

class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = ('id', 'first_name', 'last_name', 'email', 'phone', 'department', 'employee_ID')